requirements.txt
jira
python-dotenv
# Add more packages as needed for MCP automation
